+++ libraries/core/src/main/kotlin/com/bumble/appyx/core/node/Node.kt
@@ -168,8 +168,8 @@ open class Node @VisibleForTesting internal constructor(
         }
     }
 
-    override fun getLifecycle(): Lifecycle =
-        nodeLifecycle.lifecycle
+    override val lifecycle: Lifecycle
+        get() = nodeLifecycle.lifecycle
 
     override fun updateLifecycleState(state: Lifecycle.State) {
         if (lifecycle.currentState == state) return