+++ libraries/core/src/test/kotlin/com/bumble/appyx/core/lifecycle/TestLifecycle.kt
@@ -17,6 +17,7 @@ class TestLifecycle() : LifecycleOwner {
             registry.currentState = value
         }
 
-    override fun getLifecycle(): Lifecycle = registry
+    override val lifecycle: Lifecycle
+        get() = registry
 
 }